requirements.txt
django~=4.2.20
djangorestframework==3.15.2
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-environ==0.11.2
djangorestframework-simplejwt==5.4.0
pytest-django==4.8.0
requests==2.32.3
pytest~=8.3.5